Move smooth translator to util/
