+++ src/util/CMakeLists.txt
@@ -12,6 +12,7 @@ set(UTIL_SRCS
 	${CMAKE_CURRENT_SOURCE_DIR}/serialize.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/sha1.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/sha256.c
+	${CMAKE_CURRENT_SOURCE_DIR}/smoothtranslator.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/string.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/srp.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/timetaker.cpp
